Change add_user location
